Add FunctionRouterConfiguration
