+++ activiti-cloud-service-common/activiti-cloud-service-messaging-config/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,6 +1,7 @@
 org.activiti.cloud.common.messaging.config.ActivitiCloudMessagingAutoConfiguration
 org.activiti.cloud.common.messaging.config.ActivitiMessagingDestinationsAutoConfiguration
 org.activiti.cloud.common.messaging.config.ConnectorConfiguration
+org.activiti.cloud.common.messaging.config.FunctionRouterConfiguration
 org.activiti.cloud.common.messaging.config.FunctionBindingConfiguration
 org.activiti.cloud.common.messaging.config.InputBindingConfiguration
 org.activiti.cloud.common.messaging.config.OutputBindingConfiguration
